chore(circuits): reorganize the crate for better modularity
