+++ crates/circuits-core/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "mpz-circuits-core"
+version = "0.1.0-alpha.4"
+edition = "2024"
+
+[lints]
+workspace = true
+
+[lib]
+name = "mpz_circuits_core"
+
+[features]
+default = ["parse", "serde"]
+parse = ["dep:regex"]
+serde = ["dep:serde", "dep:serde_arrays", "dep:bincode"]
+
+[dependencies]
+bincode = { version = "1.3", optional = true }
+itybity = { workspace = true }
+once_cell = { workspace = true }
+rand = { workspace = true }
+regex = { workspace = true, optional = true }
+serde = { workspace = true, optional = true, features = ["derive"] }
+serde_arrays = { workspace = true, optional = true }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+aes = { workspace = true }
+sha2 = { workspace = true, features = ["compress"] }
+
+
