+++ crates/circuits-core/src/builder.rs
