+++ crates/circuits-core/src/circuit.rs
