+++ crates/circuits-core/src/components.rs
