+++ crates/ideal-vm/Cargo.toml
@@ -23,7 +23,7 @@ serio = { workspace = true }
 thiserror = { workspace = true }
 
 [dev-dependencies]
-mpz-circuits = { workspace = true }
+mpz-circuits = { workspace = true, features = ["aes"] }
 mpz-common = { workspace = true, features = ["test-utils"] }
 rstest = { workspace = true }
 tokio = { workspace = true, features = [
