+++ crates/zk-core/Cargo.toml
@@ -26,6 +26,7 @@ itybity = { workspace = true }
 zerocopy = { workspace = true }
 
 [dev-dependencies]
+mpz-circuits = { workspace = true, features = ["aes"] }
 mpz-ot-core = { workspace = true, features = ["test-utils"] }
 rand = { workspace = true }
 criterion = { workspace = true }
